5.1 SR4

+++ molap/server/5.1/Library/PaloJobs/CubeLockJob.h
@@ -99,7 +99,7 @@ public:
 			}
 
 			PArea area;
-			uint32_t numResult;
+			double numResult;
 
 			if (jobRequest->area) {
 				area = PaloJob::area(database, cube, jobRequest->area, &dims, numResult, false);
